Cleanup: Move Lines'n'Routes search components out of common
