+++ ui/src/components/common/index.ts
@@ -1,10 +1,7 @@
 export * from './DateControl';
 export * from './DateInput';
-export * from './LineTableRow';
 export * from './PageTitle';
 export * from './RedirectWithQuery';
-export * from './RouteLineTableRow';
-export * from './RouteTableRow';
 export * from './TimeRangeControl';
 export * from './versions';
 export * from './search';
