--- ui/src/components/common/search/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './mapSearchConditions';
