+++ ui/src/components/routes-and-lines/search/filters/FiltersContainer.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Row } from '../../../common/LayoutComponents';
-import { ResultSelector } from '../../../common/search/ResultSelector';
+import { ResultSelector } from '../ResultSelector';
 
 export const FiltersContainer: FC = () => (
   <Row className="my-4">
