+++ ui/src/components/routes-and-lines/search/index.ts
@@ -1,0 +1,8 @@
+export { RouteLineTableRowVariant } from './types';
+export { LinesList } from './LinesList';
+export { RoutesList } from './RoutesList';
+export { SearchContainer } from './SearchContainer';
+
+// For unit tests located in a wonky folder.
+export { LineTableRow } from './LineTableRow';
+export { RouteTableRow } from './RouteTableRow';
