+++ ui/src/components/routes-and-lines/search/types/RouteLineTableRowVariant.tsx
@@ -1,0 +1,4 @@
+export enum RouteLineTableRowVariant {
+  Timetables,
+  RoutesAndLines,
+}
