+++ ui/src/components/routes-and-lines/search/types/index.ts
@@ -1,2 +1,3 @@
+export * from './RouteLineTableRowVariant';
 export * from './RoutesAndLinesSearchFilters';
 export * from './SearchNavigationState';
